requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
Django==5.1.1
django-elasticsearch-dsl==8.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
elastic-transport==8.15.0
elasticsearch==8.15.1
elasticsearch-dsl==8.15.2
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==28.4.1
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.4
graphql-relay==3.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
promise==2.3
pytest==8.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
